w2016_05 nebula and docker

+++ scripts/cleanup.sh
@@ -1,11 +1,15 @@
-#!/bin/sh -eux
+#!/bin/bash -eux
 
-rm -f /etc/udev/rules.d/70-persistent-net.rules;
+# this has apparently become a dir
+rm -rf /etc/udev/rules.d/70-persistent-net.rules;
 yum -y clean all
 
 # cleanup centos user account (if it exists)
 if id -u centos > /dev/null 2>&1; then
-    /usr/sbin/userdel -r centos
+    # and we are not currently logged in as the centos user...
+    if [[ $(id -u -n) == 'centos' ]]; then
+        /usr/sbin/userdel -r centos
+    fi
 fi
 
 # cleanup any ssh keys nova/cloud-init may have injected
@@ -14,4 +18,7 @@ find  /home/ -maxdepth 2 -type d -name .ssh -exec rm -rf {} \;
 
 # per https://aws.amazon.com/articles/0155828273219400
 find /root/.*history /home/*/.*history -exec rm -f {} \; || true
-find /home/*/.*ssh -name "authorized_keys" –exec rm –f {} \; || true
+find /home/*/.*ssh -name "authorized_keys" -exec rm -f {} \; || true
+
+# cleanup any junk left under /tmp & /var/tmp
+find /tmp /var/tmp -mindepth 1 -maxdepth 1 -print0 | xargs -0 rm -rf
